Moved the parent resources to `parent` subfolders under the dataService and singletonDataService folders. Moved the view resources into a subfolder under the dataService folder. This makes the folder hierarchy less cluttered and easier to navigate.
