+++ source/services/dataContracts/contractLibrary/dataServiceMocks.ts
@@ -3,7 +3,7 @@
 'use strict';
 
 import { IDataService, IBaseDomainObject } from '../dataService/data.service';
-import { IParentDataService } from '../parentDataService/parentData.service';
+import { IParentDataService } from '../dataService/parent/parentData.service';
 import { ISingletonDataService } from '../singletonDataService/singletonData.service';
 
 export interface IDataServiceMock<TDataType extends IBaseDomainObject, TSearchParams> extends IDataService<TDataType, TSearchParams> {
