+++ source/services/dataContracts/dataContracts.module.ts
@@ -13,10 +13,10 @@ export var moduleName: string = 'rl.utilities.services.dataContracts';
 
 export * from './contractLibrary/contractLibrary';
 export { IDataService, IDataServiceFactory, IBaseDomainObject, DataService, factoryName as baseDataServiceFactoryName } from './dataService/data.service';
-export { IDataServiceView, IParentDataServiceView } from './dataService/dataServiceView';
-export * from './parentDataService/parentData.service';
+export { IDataServiceView, IParentDataServiceView } from './dataService/view/dataServiceView';
+export * from './dataService/parent/parentData.service';
 export { ISingletonDataService, ISingletonDataServiceFactory, SingletonDataService, factoryName as baseSingletonDataServiceFactoryName } from './singletonDataService/singletonData.service';
-export * from './parentSingletonDataService/parentSingletonData.service';
+export * from './singletonDataService/parent/parentSingletonData.service';
 export { IBaseResourceBuilder, serviceName as builderServiceName } from './resourceBuilder/resourceBuilder.service';
 export { converters, mocks };
