+++ source/services/dataContracts/dataService/view/parentDataServiceView.md
@@ -7,7 +7,7 @@ Extends [dataServiceView](./dataServiceView.md).
 
 `IChildResources`, `resourceDictionaryBuilder`
 
-See [parentDataService](../parentDataService/parentDataService.md) for details on the parent options.
+See [parentDataService](../parent/parentDataService.md) for details on the parent options.
 
 `endpoint`, `interface`, `useMock`, `logRequests`, `searchParams`, `transform`.
 
@@ -16,11 +16,11 @@ See [dataContract](../baseDataService.md) for details on the base options.
 ### Interface
 The following functions are available for consumers of the parent singleton data service.
 
-`childContracts`. See [parentDataService](../parentDataService/parentDataService.md).
+`childContracts`. See [parentDataService](../parent/parentDataService.md).
 
 `getList`, `getDetail`, `create`, `update`, `delete`, `version`.
 
-See [dataService](../dataService/dataService.md) for details on the base options.
+See [dataService](../dataService.md) for details on the base options.
 
 #### `Handling children`
 Using a parent data service view adds a little bit of complexity when defining children, since the parent itself could be a singleton or a full resource based on whether a grandparent is selected. If the child is a full resource (where the parent has a one-to-many relationship with it) we can still treat it as before. However, if the child is itself a `view`, then we have to handle it a little different. The main thing to remember here is that if a grandparent is selected, when the parent is converted to a singleton the child is likewise converted. Thus the child `view` will be treated as a singleton in either case of whether a grandparent or a direct parent is selected.
@@ -66,5 +66,5 @@ Note that in this case there are three possible urls that could be generated for
 `www.example.com/api/user/11/settings/privacy`<br/>
 You should verify that the API supports all url forms, or be careful to only use the form that is supported by the api.
 
-See [parentDataService](../parentDataService/parentDataService.md#usecases) for using a `view` against a standard parent resource.<br/>
+See [parentDataService](../parent/parentDataService.md#usecases) for using a `view` against a standard parent resource.<br/>
 See [dataServiceView](./dataServiceView.md) for details on how to define a `view`.
